requirements.txt
requests            # for testing
dnspython            # support for connecting to remote MongoDB
python-dotenv       # support for environment variables
Flask
Flask-Cors
Flask-RESTful

psycopg2-binary         # PostgreSQL DB connector
SQLAlchemy
Flask-SQLAlchemy    # Python <--> DB ORM
pylint-django        # linter help
Faker               # to install fake data
gunicorn
